requirements.txt
langgraph>=0.2.0,<0.3.0
langchain-aws>=0.1.0
langchain-core>=0.2.0,<0.3.0
langsmith>=0.1.0
PyGithub>=2.1.1
slack-bolt>=1.18.0
boto3>=1.34.0
